fix(tests): refactor _test.dart files
